requirements.txt
Django==2.2.17
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
PyMySQL==0.10.1
pytz==2020.4
redis==3.5.3
sqlparse==0.4.1
